einvestment/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.2.1
django-cors-headers==3.14.0
django-oauth-toolkit==2.3.0
django-phonenumber-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-social-oauth2==2.1.3
drf-yasg==1.21.5
ecdsa==0.18.0
geographiclib==2.0
geopy==2.3.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.5.0
MarkupSafe==2.1.2
oauthlib==3.2.2
packaging==23.1
phonenumbers==8.12.45
Pillow==9.5.0
psycopg2-binary==2.9.3
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.6.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.22
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
wrapt==1.15.0
